third_party/WebKit/LayoutTests/inspector/sass/test-edit-set-property-text.html

<html>
<head>

<link rel="stylesheet" href="resources/test-edit-set-text.css">

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="./sass-test.js"></script>
<script>

function test()
{
    var header = InspectorTest.cssModel.styleSheetHeaders().find(header => !!header.sourceMapURL)

    InspectorTest.runCSSEditTests(header, [
        function testEditPropertyName(text)
        {
            return [
                InspectorTest.createEdit(text, "color", "border")
            ];
        },

        function testEditPropertyValue(text)
        {
            return [
                InspectorTest.createEdit(text, "red", "blue")
            ];
        },

        function testEditPropertyNames(text)
        {
            return [
                InspectorTest.createEdit(text, "color", "border", 0),
                InspectorTest.createEdit(text, "color", "border", 1),
                InspectorTest.createEdit(text, "color", "border", 2),
            ];
        },

        function testEditPropertyValues(text)
        {
            return [
                InspectorTest.createEdit(text, "red", "blue", 0),
                InspectorTest.createEdit(text, "red", "blue", 1),
                InspectorTest.createEdit(text, "red", "blue", 2)
            ];
        },

        function testConflictingEdits(text)
        {
            return [
                InspectorTest.createEdit(text, "red", "green", 0),
                InspectorTest.createEdit(text, "red", "magenta", 1),
                InspectorTest.createEdit(text, "red", "yellow", 2)
            ];
        },
    ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verify that mapping is not valid for misaligned sources.</p>
</body>
</html>
