third_party/WebKit/LayoutTests/inspector/sass/test-edit-set-property-text-expected.txt

Verify that mapping is not valid for misaligned sources.

INITIAL MODELS
===== test-edit-set-text.css =====
.box1.{
..color:.red;
..margin:.10px;.}

.box2.{
..color:.red;
..margin:.10px;.}

.box3.{
..color:.red;
..margin:.10px;.}

body.{
..font-family:."arial";.}

/*#.sourceMappingURL=test-edit-set-text.css.map.*/

===== test-edit-set-text.scss =====
$color:.red;

@for.$i.from.1.through.3.{
.....box#{$i}.{
........color:.$color;
........margin:.10px;
....}
}

body.{
....font-family:."arial";
}


------------------------- TEST: testEditPropertyName -------------------------

Edits:
    {1, 2, 1, 7} 'color' => 'border'
===== test-edit-set-text.css =====
    .box1.{
    ..border:.red;
    ..margin:.10px;.}
    
    .box2.{
    ..border:.red;
    ..margin:.10px;.}
    
    .box3.{
    ..border:.red;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-set-text.css.map.*/
    
===== test-edit-set-text.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........border:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

------------------------ TEST: testEditPropertyValue -------------------------

Edits:
    {1, 9, 1, 12} 'red' => 'blue'
===== test-edit-set-text.css =====
    .box1.{
    ..color:.blue;
    ..margin:.10px;.}
    
    .box2.{
    ..color:.blue;
    ..margin:.10px;.}
    
    .box3.{
    ..color:.blue;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-set-text.css.map.*/
    
===== test-edit-set-text.scss =====
    $color:.blue;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........color:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

------------------------ TEST: testEditPropertyNames -------------------------

Edits:
    {1, 2, 1, 7} 'color' => 'border'
    {5, 2, 5, 7} 'color' => 'border'
    {9, 2, 9, 7} 'color' => 'border'
===== test-edit-set-text.css =====
    .box1.{
    ..border:.red;
    ..margin:.10px;.}
    
    .box2.{
    ..border:.red;
    ..margin:.10px;.}
    
    .box3.{
    ..border:.red;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-set-text.css.map.*/
    
===== test-edit-set-text.scss =====
    $color:.red;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........border:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

------------------------ TEST: testEditPropertyValues ------------------------

Edits:
    {1, 9, 1, 12} 'red' => 'blue'
    {5, 9, 5, 12} 'red' => 'blue'
    {9, 9, 9, 12} 'red' => 'blue'
===== test-edit-set-text.css =====
    .box1.{
    ..color:.blue;
    ..margin:.10px;.}
    
    .box2.{
    ..color:.blue;
    ..margin:.10px;.}
    
    .box3.{
    ..color:.blue;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-set-text.css.map.*/
    
===== test-edit-set-text.scss =====
    $color:.blue;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........color:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    

------------------------- TEST: testConflictingEdits -------------------------

Edits:
    {1, 9, 1, 12} 'red' => 'green'
    {5, 9, 5, 12} 'red' => 'magenta'
    {9, 9, 9, 12} 'red' => 'yellow'
===== test-edit-set-text.css =====
    .box1.{
    ..color:.green;
    ..margin:.10px;.}
    
    .box2.{
    ..color:.green;
    ..margin:.10px;.}
    
    .box3.{
    ..color:.green;
    ..margin:.10px;.}
    
    body.{
    ..font-family:."arial";.}
    
    /*#.sourceMappingURL=test-edit-set-text.css.map.*/
    
===== test-edit-set-text.scss =====
    $color:.green;
    
    @for.$i.from.1.through.3.{
    .....box#{$i}.{
    ........color:.$color;
    ........margin:.10px;
    ....}
    }
    
    body.{
    ....font-family:."arial";
    }
    
